+++ philo/srcs/utils.c
@@ -6,7 +6,7 @@
 /*   By: jthuysba <jthuysba@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/17 11:59:38 by jules             #+#    #+#             */
-/*   Updated: 2023/03/13 16:46:54 by jthuysba         ###   ########.fr       */
+/*   Updated: 2023/03/13 20:58:38 by jthuysba         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,11 +21,11 @@ int	ft_isdigit(int c)
 }
 
 //Converti un num str en int
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
-	int	    i;
-	int		res;
-	int		sign;
+	long	    i;
+	long		res;
+	long		sign;
 
 	i = 0;
 	res = 0;
